Revert workspace to webdav. Use overlayfs for rrun

+++ gwvolman/tasks.py
@@ -70,7 +70,7 @@ def create_volume(self, instance_id):
     _mount_bind(mountpoint, 'home', user)
 
     if ENABLE_WORKSPACES:
-        _mount_overlayfs(mountpoint, 'workspace', tale)
+        _mount_girderfs(mountpoint, 'workspace', 'wt_work', tale['_id'], api_key)
         _mount_girderfs(mountpoint, 'versions', 'wt_versions', tale['_id'], api_key, hostns=True)
         _mount_girderfs(mountpoint, 'runs', 'wt_runs', tale['_id'], api_key, hostns=True)
 
@@ -276,8 +276,6 @@ def remove_volume(self, instanceId):
     for suffix in MOUNTPOINTS:
         dest = os.path.join(containerInfo['mountPoint'], suffix)
         logging.info("Unmounting %s", dest)
-        if suffix == "workspace":
-            _umount_overlayfs(dest)
         subprocess.call("umount %s" % dest, shell=True)
 
     logging.info("Unmounting licenses")
@@ -631,18 +629,28 @@ def _mount_girderfs(mountpoint, directory, fs_type, obj_id, api_key, hostns=Fals
 def _mount_overlayfs(mountpoint, res_type, obj):
     if res_type == "workspace":
         lower_dir = f"{VOLUMES_ROOT}/workspaces/{obj['_id'][0]}/{obj['_id']}"
+        upper_dir = f"{lower_dir}_upper"
+        work_dir = f"{lower_dir}_work"
+    elif res_type == "run":
+        lower_dir = (
+            f"{VOLUMES_ROOT}/versions/{obj['taleId'][0:2]}/{obj['taleId']}/"
+            f"{obj['versionId']}/workspace"
+        )
+        upper_dir = (
+            f"{VOLUMES_ROOT}/runs/{obj['taleId'][0:2]}/{obj['taleId']}/"
+            f"{obj['runId']}/workspace"
+        )
+        work_dir = f"{upper_dir}_"
+        res_type = "workspace"
     else:
         raise ValueError(f"Unknown bind type {res_type}")
-
-    upper_dir = f"{lower_dir}_upper"
-    work_dir = f"{lower_dir}_work"
 
     options = (
         f"lowerdir={lower_dir},upperdir={upper_dir},workdir={work_dir}"
         ",index=off,xino=off,redirect_dir=off,metacopy=off"
     )
-    os.makedirs(HOSTDIR + upper_dir)
-    os.makedirs(HOSTDIR + work_dir)
+    os.makedirs(HOSTDIR + upper_dir, exist_ok=True)
+    os.makedirs(HOSTDIR + work_dir, exist_ok=True)
     cmd = f'mount -t overlay overlay -o "{options}" "{mountpoint}/{res_type}"'
     logging.info("Calling: %s", cmd)
     subprocess.check_call(cmd, shell=True)
@@ -650,7 +658,7 @@ def _mount_overlayfs(mountpoint, res_type, obj):
     print(f"Mounted {lower_dir} to {mountpoint}/{res_type}")
 
 
-def _umount_overlayfs(mountpoint):
+def _umount_overlayfs(mountpoint, sync=False, rm_upper=False):
     # Get required paths
     mount_details = subprocess.check_output(f"mount | grep {mountpoint}", shell=True).decode()
     lowerdir = re.search("lowerdir=(.+?),", mount_details).group(1)
@@ -658,14 +666,16 @@ def _umount_overlayfs(mountpoint):
     workdir = re.search("workdir=(.+?),", mount_details).group(1)
     # Umount overlayfs
     subprocess.check_call(f"umount {mountpoint}", shell=True)
-    overlay_output = subprocess.check_output(
-        f"overlay merge --lowerdir={HOSTDIR + lowerdir} --upperdir={HOSTDIR + upperdir}",
-        shell=True
-    ).decode()
-    sync_script = re.search("The script (.+?) is created.", overlay_output).group(1)
-    subprocess.check_call(f"./{sync_script}", shell=True)
-    os.remove(sync_script)
-    shutil.rmtree(f"{HOSTDIR + upperdir}", ignore_errors=True)
+    if sync:
+        overlay_output = subprocess.check_output(
+            f"overlay merge --lowerdir={HOSTDIR + lowerdir} --upperdir={HOSTDIR + upperdir}",
+            shell=True
+        ).decode()
+        sync_script = re.search("The script (.+?) is created.", overlay_output).group(1)
+        subprocess.check_call(f"./{sync_script}", shell=True)
+        os.remove(sync_script)
+    if rm_upper:
+        shutil.rmtree(f"{HOSTDIR + upperdir}", ignore_errors=True)
     shutil.rmtree(f"{HOSTDIR + workdir}", ignore_errors=True)
 
 
@@ -783,7 +793,11 @@ def recorded_run(self, run_id, tale_id, entrypoint):
 
     if session['_id'] is not None:
         _mount_girderfs(mountpoint, 'data', 'wt_dms', session['_id'], api_key, hostns=True)
-    _mount_bind(mountpoint, "run", {"runId": run["_id"], "taleId": tale["_id"]})
+    _mount_overlayfs(
+        mountpoint,
+        "run",
+        {"runId": run["_id"], "taleId": tale["_id"], "versionId": run["runVersionId"]},
+    )
     state.fs_mounted = mountpoint
 
     # Build the image for the run
@@ -861,10 +875,13 @@ class RecordedRunCleaner:
     def cleanup(self, canceled=True):
 
         if self.fs_mounted:
-            for suffix in ["data", "workspace"]:
-                dest = os.path.join(self.fs_mounted, suffix)
-                logging.info("Unmounting %s", dest)
-                subprocess.call("umount %s" % dest, shell=True)
+            dest = os.path.join(self.fs_mounted, "data")
+            logging.info("Unmounting %s", dest)
+            subprocess.call("umount %s" % dest, shell=True)
+
+            dest = os.path.join(self.fs_mounted, "workspace")
+            logging.info("Unmounting %s", dest)
+            _umount_overlayfs(dest, sync=False, rm_upper=False)
             self.fs_mounted = None
 
         if self.session_created:
